document react-navigation-header-buttons
So I wrote this package for rendering header buttons that should make people's life easier: https://github.com/vonovak/react-navigation-header-buttons (demo: https://expo.io/@vonovak/navbar-buttons-demo).

This place in the docs is ideal for it to appear, but I'm not sure if it's a good idea to link to some random guy's git repo in the middle of the page (it doesn't seem to be a good precedent to me).

I was thinking we may create a new page with community contributions and collect such packages there? (I have one more that I'd like to share since it may be useful: https://github.com/vonovak/react-navigation-props-mapper).
Alternatively, it may be moved to the react-navigation org and I could maintain it. Then it would be okay to link to it, but again cannot be applied to all community-developed packages.

Thanks for feedback!